requirements.txt
dnspython~=2.1.0
iteration_utilities
slack_sdk
